(实体ubuntu环境)requirements.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==1.8.2
django-redis==4.10.0
django-tinymce==2.6.0
fdfs-client-py==1.2.6
idna==2.8
itsdangerous==1.1.0
kombu==4.2.2.post1
mutagen==1.42.0
PyMySQL==0.9.3
pytz==2018.9
redis==3.0.1
requests==2.21.0
urllib3==1.24.1
uWSGI==2.0.17.1
vine==1.2.0
xlwt==1.3.0
